requirements.txt
openai
icecream
supabase
streamlit
python-dotenv
langchain
langchain_experimental
langchain_openai
unstructured
tiktoken
simsimd